clinical_trial/requirements.txt
aniso8601==9.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
distlib==0.3.2
et-xmlfile==1.1.0
filelock==3.0.12
Flask-Cors==3.0.10
flask-restx==0.5.1
greenlet==1.1.1
idna==3.2
jsonschema==3.2.0
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
Pillow==8.3.2
platformdirs==2.3.0
pycparser==2.20
PyMySQL==1.0.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.7
virtualenv==20.7.2
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
uWSGI==2.0.19.1
Werkzeug==1.0.1
